Firefox support for un-blurred previews
As-written, preview images were un-interpolated (and therefore un-blurred) in Chrome, Edge, Safari mobile, and Chrome mobile, but broken in Internet Explorer and Firefox (these are the browsers I was able to test).

Image-rendering is a new property so the browsers don't all use the same algorithms for the same values.  IE doesn't support the property at all, but Firefox does, and using "crisp-edges" makes it show the same behaviour that "pixelated" does on other browsers.